refactor: break resources results into smaller component within larger container
